\name{Fattorini}
\alias{Fattorini}
\docType{data}
\title{
Australian bird abudances.
}
\description{
The data sets consists of the names and abundances of 31 Australian birds.
}
\usage{data("Fattorini")}
\format{
  A data frame with 31 observations on the following 2 variables.
  \describe{
    \item{\code{Species}}{a character vector}
    \item{\code{Abundance}}{a numeric vector}
  }
}
\references{
Fattorini, S. (2005) A simple method to fit geometric series and broken stick models in community ecology and island biogeography.
Acta Oecologica 28: pp. 199-205.
}
\examples{
data(Fattorini)
}
\keyword{datasets}
