\name{Ganeshaiah}
\alias{Ganeshaiah}
\docType{data}
\title{Indian dung beetles from Ganeshaiah et al. (1997)}
\description{
The data sets consists of the names and abundances of 16 dung beetles
}
\usage{data(Ganeshaiah)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Species}}{a character vector}
    \item{\code{Abundance}}{a numeric vector}
  }
}
\references{
Ganeshaiah, K.N., Chandrashekara, K. and Kuma A.R.V. (1997) Avalanche index: a new measure of biodiversity based on biological heterogeneity of the communities. Current Science 73, pp. 128-133.

Magurran, A.E. (2004) Measuring biological diversity, Blackwell Publishing, Oxford, UK.
}
\examples{
data(Ganeshaiah)
}
\keyword{datasets}
