\name{Mehrabi}
\alias{Mehrabi}
\docType{data}
\title{
Costa Rica dung beetle counts from Mehrabi et al. (2014)
}
\description{
The data sets consists of the names and total abundances of 31 dung beetles along 16 transects (A, B, ... P). Transects A, C, ..., O used randomly placed traps (control), whereas transects B, D, ..., P used microhabitat standardized traps (treatment).  
}
\usage{data("Mehrabi")}
\format{
  A data frame with 31 observations on the following 16 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\code{J}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{O}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
  }
}
\references{
Mehrabi, Z., Slade, E.M., Solis, A. and Mann, D.J. (2014) The Importance of Microhabitat for Biodiversity Sampling (2014)
PLoS ONE 9(12) e114015.  doi 10.1371/journal.pone.0114015
}
\examples{
data(Mehrabi)
}
\keyword{datasets}
